Move patch_regions out of models/, into media/
`media/` is the lower layer in the embed/sort stack, so importing types
from `models/` was backwards. Six image embedders and `media/embedder.py`
were reaching up into `vtsearch.models.patch_regions` for `PatchEmbedOutput`
and friends; `media/embedder.py` also pulled `ensure_torch_configured`
out of `models/loader.py` for the same reason.

- `vtsearch/models/patch_regions.py` → `vtsearch/media/patch_embed.py`
  (verbatim move; the module never depended on anything in `models/`).
- `ensure_torch_configured` moves to a new `vtsearch/media/torch_setup.py`;
  `models/loader.py` re-exports it so its existing callers (and tests
  patching the loader namespace) keep working.
- All 7 patch-regions import sites updated, plus the two HAC sweep
  scripts and the patch-embedder test.
- Docstring/path references in `vtsearch/config.py`,
  `docs/plans/patch-embedder.md`, and `vtsearch/media/embedder.py`
  point at the new module.

Also drops three unused `MagicMock` imports flagged by ruff in
`tests/downloads/` (pre-existing on dev; fixed because lint surfaced
them).
