+++ tests/downloads/test_bbc_news_download.py
@@ -2,7 +2,7 @@
 
 import zipfile
 from pathlib import Path
-from unittest.mock import MagicMock, patch
+from unittest.mock import patch
 
 import pytest
