+++ tests/downloads/test_imdb_download.py
@@ -2,7 +2,7 @@
 
 import tarfile
 from pathlib import Path
-from unittest.mock import MagicMock, patch
+from unittest.mock import patch
 
 
 # ---------------------------------------------------------------------------
