+++ vtsearch/media/image/embedder_eupe_patch.py
@@ -21,14 +21,14 @@ from __future__ import annotations
 from typing import Optional
 
 from vtsearch.media.image._eupe_shared import _EupeBase
-from vtsearch.models.patch_regions import PatchEmbedOutput
+from vtsearch.media.patch_embed import PatchEmbedOutput
 
 
 class ImageEupePatchEmbedder(_EupeBase):
     """Embeds images using facebookresearch/EUPE ViT-B/16 with CLS + patch regions.
 
     Output dimension: 768.  Per-image side-channel: HAC region tree +
-    14 × 14 patch grid (see :func:`vtsearch.models.patch_regions`).
+    14 × 14 patch grid (see :func:`vtsearch.media.patch_embed`).
     """
 
     @property
